requirements.txt
requests
discord.py
pychallonge
wikitextparser